major update, gallery,stories,chat

+++ app/Http/Controllers/Escorts.php
@@ -10,16 +10,17 @@ class Escorts extends Controller
     public function kinks(Request $request)
     {
         try {
-            $kinks = User::whenHas('tags', function ($query) use ($request) {
-                $query->whereJsonContains('tags', $request->tags);
-            })->whenHas('fees', function ($query) use ($request) {
-                $fee = explode('-', $request->fees);
-                $query->whereBetween('amount', [$fee[0], $fee[1]]);
-            })->whenHas('location', function($query) use ($request) {
-                $query->where('location', $request->location);
-            })->where('is_escort', true)->inRandomOrder()->get();
-        
-            return get_success_response($kinks);
+            $query = User::where('is_escort', true)->inRandomOrder()->get();
+            // $query = $request->whenHas('tags', function ($query) use ($request) {
+            //     $query->whereJsonContains('tags', $request->tags);
+            // })->whenHas('fees', function ($query) use ($request) {
+            //     $fee = explode('-', $request->fees);
+            //     $query->whereBetween('amount', [$fee[0], $fee[1]]);
+            // })->whenHas('location', function ($query) use ($request) {
+            //     $query->where('location', $request->location);
+            // });
+
+            return get_success_response($query);
         } catch (\Throwable $th) {
             return get_error_response(['error' =>  $th->getMessage()]);
         }
@@ -28,7 +29,7 @@ class Escorts extends Controller
     public function kink(Request $request, $id)
     {
         try {
-            $kink = User::where('id', $id)->with('messages',  'orders')->findorfail();
+            $kink = User::with('messages', 'orders')->find($id);
             return get_success_response($kink);
         } catch (\Throwable $th) {
             return get_error_response(['error' =>  $th->getMessage()]);
@@ -38,8 +39,11 @@ class Escorts extends Controller
     public function destroy($id)
     {
         try {
-            $record =  User::whereUserId(auth()->id())->whereId($id)->findorfail();
-            if($record->delete())
+            if((int)auth()->id() !== (int)$id) {
+                return get_error_response(['error' =>  'You can only delete your own account']);
+            }
+            $record =  User::whereId($id)->first();
+            if ($record->delete())
                 return get_success_response(['message' => 'Record deleted successfully']);
             return get_error_response(['error' =>  'Unable to delete, please contact support']);
         } catch (\Throwable $th) {
